Fastapi/requirements.txt
absl-py
annotated-types
anyio
astunparse
cachetools
certifi
charset-normalizer
fastapi
flatbuffers
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
keras
libclang
Markdown
MarkupSafe
ml-dtypes
numpy
oauthlib
opt-einsum
packaging
Pillow
protobuf
pyasn1
pyasn1-modules
pydantic
pydantic_core
python-multipart
requests
requests-oauthlib
rsa
six
sniffio
starlette
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
typing_extensions
urllib3
Werkzeug
wrapt
gunicorn
fastapi[all]